+++ sdk/operationalinsights/arm-operationalinsights/samples/v11/javascript/workspacePurgeGetPurgeStatusSample.js
@@ -1,0 +1,48 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { OperationalInsightsManagementClient } = require("@azure/arm-operationalinsights");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to gets status of an ongoing purge operation.
+ *
+ * @summary gets status of an ongoing purge operation.
+ * x-ms-original-file: 2026-03-01/WorkspacesPurgeLakeDataOperation.json
+ */
+async function workspacePurgeLakeDataOperation() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
+  const client = new OperationalInsightsManagementClient(credential, subscriptionId);
+  const result = await client.workspacePurge.getPurgeStatus(
+    "OIAutoRest5123",
+    "aztest5048",
+    "purgelakedata-7d7cf277-9113-4ab3-8359-d0364b74d01d",
+  );
+  console.log(result);
+}
+
+/**
+ * This sample demonstrates how to gets status of an ongoing purge operation.
+ *
+ * @summary gets status of an ongoing purge operation.
+ * x-ms-original-file: 2026-03-01/WorkspacesPurgeOperation.json
+ */
+async function workspacePurgeOperation() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
+  const client = new OperationalInsightsManagementClient(credential, subscriptionId);
+  const result = await client.workspacePurge.getPurgeStatus(
+    "OIAutoRest5123",
+    "aztest5048",
+    "purge-970318e7-b859-4edb-8903-83b1b54d0b74",
+  );
+  console.log(result);
+}
+
+async function main() {
+  await workspacePurgeLakeDataOperation();
+  await workspacePurgeOperation();
+}
+
+main().catch(console.error);