+++ sdk/operationalinsights/arm-operationalinsights/src/api/deletedWorkspaces/options.ts
@@ -1,10 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OperationOptions } from "@azure-rest/core-client";
+import type { OperationOptions } from "@azure-rest/core-client";
 
 /** Optional parameters. */
 export interface DeletedWorkspacesListByResourceGroupOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface DeletedWorkspacesListOptionalParams extends OperationOptions {}