+++ sdk/operationalinsights/arm-operationalinsights/src/api/intelligencePacks/options.ts
@@ -1,13 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OperationOptions } from "@azure-rest/core-client";
+import type { OperationOptions } from "@azure-rest/core-client";
 
 /** Optional parameters. */
 export interface IntelligencePacksListOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface IntelligencePacksEnableOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface IntelligencePacksDisableOptionalParams extends OperationOptions {}