+++ sdk/operationalinsights/arm-operationalinsights/src/api/storageInsightConfigs/options.ts
@@ -1,16 +1,13 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OperationOptions } from "@azure-rest/core-client";
+import type { OperationOptions } from "@azure-rest/core-client";
 
 /** Optional parameters. */
 export interface StorageInsightConfigsListByWorkspaceOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface StorageInsightConfigsDeleteOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface StorageInsightConfigsCreateOrUpdateOptionalParams extends OperationOptions {}
-
 /** Optional parameters. */
 export interface StorageInsightConfigsGetOptionalParams extends OperationOptions {}