+++ sdk/operationalinsights/arm-operationalinsights/src/classic/clusters/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OperationalInsightsManagementContext } from "../../api/operationalInsightsManagementContext.js";
+import type { OperationalInsightsManagementContext } from "../../api/operationalInsightsManagementContext.js";
 import {
   list,
   listByResourceGroup,
@@ -10,7 +10,7 @@ import {
   createOrUpdate,
   get,
 } from "../../api/clusters/operations.js";
-import {
+import type {
   ClustersListOptionalParams,
   ClustersListByResourceGroupOptionalParams,
   ClustersDeleteOptionalParams,
@@ -18,10 +18,11 @@ import {
   ClustersCreateOrUpdateOptionalParams,
   ClustersGetOptionalParams,
 } from "../../api/clusters/options.js";
-import { Cluster, ClusterPatch } from "../../models/models.js";
-import { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
-import { SimplePollerLike, getSimplePoller } from "../../static-helpers/simplePollerHelpers.js";
-import { PollerLike, OperationState } from "@azure/core-lro";
+import type { Cluster, ClusterPatch } from "../../models/models.js";
+import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
+import type { SimplePollerLike } from "../../static-helpers/simplePollerHelpers.js";
+import { getSimplePoller } from "../../static-helpers/simplePollerHelpers.js";
+import type { PollerLike, OperationState } from "@azure/core-lro";
 
 /** Interface representing a Clusters operations. */
 export interface ClustersOperations {
@@ -99,7 +100,6 @@ export interface ClustersOperations {
     options?: ClustersGetOptionalParams,
   ) => Promise<Cluster>;
 }
-
 function _getClusters(context: OperationalInsightsManagementContext) {
   return {
     list: (options?: ClustersListOptionalParams) => list(context, options),
@@ -180,7 +180,6 @@ function _getClusters(context: OperationalInsightsManagementContext) {
       get(context, resourceGroupName, clusterName, options),
   };
 }
-
 export function _getClustersOperations(
   context: OperationalInsightsManagementContext,
 ): ClustersOperations {