+++ sdk/operationalinsights/arm-operationalinsights/src/classic/linkedStorageAccounts/index.ts
@@ -1,21 +1,21 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { OperationalInsightsManagementContext } from "../../api/operationalInsightsManagementContext.js";
+import type { OperationalInsightsManagementContext } from "../../api/operationalInsightsManagementContext.js";
 import {
   listByWorkspace,
   $delete,
   createOrUpdate,
   get,
 } from "../../api/linkedStorageAccounts/operations.js";
-import {
+import type {
   LinkedStorageAccountsListByWorkspaceOptionalParams,
   LinkedStorageAccountsDeleteOptionalParams,
   LinkedStorageAccountsCreateOrUpdateOptionalParams,
   LinkedStorageAccountsGetOptionalParams,
 } from "../../api/linkedStorageAccounts/options.js";
-import { LinkedStorageAccountsResource, DataSourceType } from "../../models/models.js";
-import { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
+import type { LinkedStorageAccountsResource, DataSourceType } from "../../models/models.js";
+import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 
 /** Interface representing a LinkedStorageAccounts operations. */
 export interface LinkedStorageAccountsOperations {
@@ -48,7 +48,6 @@ export interface LinkedStorageAccountsOperations {
     options?: LinkedStorageAccountsGetOptionalParams,
   ) => Promise<LinkedStorageAccountsResource>;
 }
-
 function _getLinkedStorageAccounts(context: OperationalInsightsManagementContext) {
   return {
     listByWorkspace: (
@@ -85,7 +84,6 @@ function _getLinkedStorageAccounts(context: OperationalInsightsManagementContext
     ) => get(context, resourceGroupName, workspaceName, dataSourceType, options),
   };
 }
-
 export function _getLinkedStorageAccountsOperations(
   context: OperationalInsightsManagementContext,
 ): LinkedStorageAccountsOperations {