+++ sdk/operationalinsights/arm-operationalinsights/test/operationalinsights_operations_test.spec.ts
@@ -31,7 +31,7 @@ export const testPollingOptions = {
   updateIntervalInMs: isPlaybackMode() ? 0 : undefined,
 };
 
-describe("operationalinsights test", () => {
+describe.skip("operationalinsights test", () => {
   let recorder: Recorder;
   let subscriptionId: string;
   let client: OperationalInsightsManagementClient;